Moved most of the cookie and network interaction to the injected page
When "keep me logged in" was not used during log in, the session cookie
is not accessible in the global page. By moving the network request to
fetch the CSRF token to the injected/content page (popups.js) this
works now.

So, the structure is more like this:
* global page: for UI + I18N
* injected page: network requests + DOM handling of current wiki page

Between those messages have to be dispatched. Care has to be taken that
only simple, cloneable objects are passed with the message as message
data.

Bug: T191998