Move remaining network calls to injected script
We can't omit cookies in the i18n message fetch calls because this means
disregarding the user's language preferences. But sending them from the
popup script results in Safari ending the session (at least on some
machines...).  This moves the i18n message fetches over to the injected
script, so that the session cookie issue is resolved once and for all.