Fix bug with button enabled state when using back or forward buttons
Fixes a bug in which the page action could be inappropriately shown for
non-Wikimedia pages when navigating back or forward from a Wikimedia page.

This update makes better use of Safari lifecycle events to disable the
button before every navigation event, and enable it only when the expected
conditions are met.

Bug: https://phabricator.wikimedia.org/T197825